hard code email route
